requirements.txt
alembic==1.6.5
aniso8601==9.0.1
autopep8==1.5.4
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click 
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
flask-paginate==0.8.1
Flask-RESTful==0.3.9
Flask-Scss==0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata
itsdangerous 
Jinja2 
Mako==1.1.4
MarkupSafe
mysqlclient==2.0.3
oauthlib==3.1.1
Pillow==8.3.1
pycodestyle==2.6.0
pycparser==2.20
pyScss==1.3.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
urllib3==1.26.6
URLObject==2.4.3
Werkzeug
wincertstore==0.2
WTForms==2.3.3
zipp
